New GWT Login architecture. The basic AuthenticationProcessingFilter class will either forward us to the 'savedRequest' page on login, or it will send us to the defaultTargetURL. This is ok normally, but we'd like to be able to post a from from GWT to login. When we do that, it's inefficeint (and plays havoc with hosted mode) to return the entire defaultTargetURL (in this case the MyPage). Let's give ourselves the ability to forward to a different page if there's a get parameter in the URL. There we can simply return 'OK' so that GWT knows it worked. See the appropriate mapping in SimpleAnnotatedController.java.
For logging in, callers now need to specify a URL that we can use to perform possible
login via forwarding. We can't do OpenID logins with a simple
asynchronous form POST (maybe with Comet?) because the user might
need to go offsite to type in their URL. When they come back from
hat, we need to have a URL that will bring us back to the desired
state of the application.

Of course for Username/Password logins, this URL will not be used,
we'll just do the asyncronous FormPost.