Making test user supervisor in init.sql Better error page for 404 and 500 errors Spring 2.5.2
Mac launch fixes
OSX hosted mode throws errors when port 8888 tries to read from port 8080 (which happens during our hosted mode login). This is annoying, but we can avoid this by running in -noserver mode so that hosted mode is on port 8080 just like the jetty process.

run_jetty_gwt requires that mvn package has been run so that WEB-INF/web.xml is in the target directory. Added this line to out run_jetty_gwt script so that this happens automatically.